+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## Next
 ### Added
 * duplicate the build_image function and make it asynchronous
+### Fixed
+* Fix the way to import exception from Aerospike
 ## 0.8.5
 ### Fixed
 * `RunMixin` and `AsyncRunMixin` no longer subclass `ContainerService`